create H2 database, added patinet service

+++ src/main/java/com/example/medicalclinic/dto/PatientDTO.java
@@ -10,7 +10,7 @@ public class PatientDTO {
     @NotEmpty
     private String firstName;
     private String surname;
-    private int pesel;
+    private long pesel;
     private int phoneNumber;
     private String comment;
 }
